# mcg/docs/openapi.template.yaml
...
openapi: 3.0.1
info:
  title: Metrics Config Generator
  description: Generate MetricsConfig for the SDV Telemetry Service
  contact:
    name: Android SDV Telemetry
    email: aaos-sdv-telemetry-external@google.com
  version: "<<<REPLACED DURING BUILD>>>"
servers:
- url: '{url}'
  variables:
    url:
      default: "/"
      description: The base URL of MCG (use `/` for the current URL).
paths:
...
  /api/v1/vs/:
    get:
      tags:
      - v1 — Vehicle Signal Catalog metadata
      description: List all Vehicle Signal Catalog metadata
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/signal_versions.VSCatalogVersionsMetadata"
        "400":
          description: Cache configuration errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors.JsonErrorResponse"
    post:
      tags:
      - v1 — Vehicle Signal Catalogs
      description: Create new or update existing Vehicle Signal Catalog
      requestBody:
        description: VSCatalogVersion
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/signal_versions.VSCatalogVersion"
        required: false
      responses:
        "200":
          description: Vehicle Signal Catalog added/updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/signal_versions.VSCatalogVersionMetadata"
        "400":
          description: Cache configuration or Invalid Request errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors.JsonErrorResponse"
      x-codegen-request-body-name: body
  /api/v1/vs/{version}:
    delete:
      tags:
      - v1 — Vehicle Signal Catalogs
      description: Delete Vehicle Signal Catalog
      parameters:
      - name: version
        in: path
        description: Vendor-defined identifier of the Vehicle Signal Catalog to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Vehicle Signal Catalog deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/signal_versions.VSCatalogVersionMetadata"
        "400":
          description: Cache configuration or Invalid Request errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors.JsonErrorResponse"
        "404":
          description: Vehicle Signal Catalog not found
          content:
            application/json:
              schema:
                type: object
...
  /api/v2/vs/:
    get:
      tags:
      - v2 — Vehicle Signal Catalog metadata
      description: List all Vehicle Signal Catalog metadata
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/signal_versions.VSCatalogVersionsMetadata"
        "400":
          description: Cache configuration errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors.JsonErrorResponse"
    post:
      tags:
      - v2 — Vehicle Signal Catalogs
      description: Create new or update existing Vehicle Signal Catalog
      requestBody:
        description: VSCatalogVersion
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/signal_versions.VSCatalogVersion"
        required: false
      responses:
        "200":
          description: Vehicle Signal Catalog added/updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/signal_versions.VSCatalogVersionMetadata"
        "400":
          description: Cache configuration or Invalid Request errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors.JsonErrorResponse"
      x-codegen-request-body-name: body
  /api/v2/vs/{version}:
    delete:
      tags:
      - v2 — Vehicle Signal Catalogs
      description: Delete Vehicle Signal Catalog
      parameters:
      - name: version
        in: path
        description: Vendor-defined identifier of the Vehicle Signal Catalog to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Vehicle Signal Catalog deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/signal_versions.VSCatalogVersionMetadata"
        "400":
          description: Cache configuration or Invalid Request errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors.JsonErrorResponse"
        "404":
          description: Vehicle Signal Catalog not found
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
    errors.JsonErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/errors.JsonErrorResponse_error"
...
    signal_versions.VSCatalogVersion:
      required:
      - vehicle_signals
      - version
      type: object
      properties:
        vehicle_signals:
          type: string
          description: A `google.protobuf.FileDescriptorSet` protobuf message (base64-encoded)
            containing the Vehicle Signals
          format: base64
          example: Cq4BChxhZnRlcl9ldmVudF9jb2xsZWN0aW9uLnByb3RvEhxnb29nbGUuc2R2LnRlbGVtZXRyeV9zYW1wbGVzIkYKCEV2ZW50TXNnEhQKBWV2ZW50GAEgASgFUgVldmVudBIkCg5pc19maXJzdF9ldmVudBgCIAEoCFIMaXNGaXJzdEV2ZW50IiAKCFNwZWVkTXNnEhQKBXNwZWVkGAEgASgFUgVzcGVlZGIGcHJvdG8z
        version:
          minLength: 1
          type: string
          description: Vendor-defined identifier of the catalog
          example: v1
      description: Vehicle Signal Catalog
    signal_versions.VSCatalogVersionMetadata:
      required:
      - version
      type: object
      properties:
        version:
          minLength: 1
          type: string
          description: Vendor-defined identifier of the catalog
          example: v1
      description: Vehicle Signal Catalog metadata
    signal_versions.VSCatalogVersionsMetadata:
      required:
      - versions
      type: object
      properties:
        versions:
          type: array
          description: List of Vehicle Signal Catalog metadata
          items:
            $ref: "#/components/schemas/signal_versions.VSCatalogVersionMetadata"
      description: A list of Vehicle Signal Catalog metadata
    errors.JsonErrorResponse_error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 403
        details:
          type: array
          description: Array of objects containing further error details identified by their @type member
          items:
            type: object
            properties: {}
        message:
          type: string
          description: Developer-facing message in English
        status:
          type: string
          description: Google RPC canonical status code
          enum:
          - OK
          - CANCELLED
          - UNKNOWN
          - INVALID_ARGUMENT
          - DEADLINE_EXCEEDED
          - NOT_FOUND
          - ALREADY_EXISTS
          - PERMISSION_DENIED
          - UNAUTHENTICATED
          - RESOURCE_EXHAUSTED
          - FAILED_PRECONDITION
          - ABORTED
          - OUT_OF_RANGE
          - UNIMPLEMENTED
          - INTERNAL
          - UNAVAILABLE
          - DATA_LOSS
      example:
        code: 400
        status: INVALID_ARGUMENT
        message: Request validation failed
        details:
          - '@type': type.googleapis.com/google.rpc.BadRequest
            fieldViolations:
              - field: fieldName
                description: error description